+++ packages/iov-stream/src/defaultvalueproducer.spec.ts
@@ -49,6 +49,24 @@ describe("DefaultValueProducer", () => {
     producer.update(45);
   });
 
+  it("can send errors", done => {
+    const producer = new DefaultValueProducer(42);
+    const stream = Stream.createWithMemory(producer);
+
+    stream.addListener({
+      error: error => {
+        expect(error).toEqual("oh no :(");
+        done();
+      },
+      complete: () => done.fail("Stream must not complete sucessfully"),
+    });
+
+    producer.update(1);
+    producer.update(2);
+    producer.update(3);
+    producer.error("oh no :(");
+  });
+
   it("calls callbacks", async () => {
     // tslint:disable-next-line:readonly-array
     const producerActive: boolean[] = [];